+++ BCI_V3/Modules/Core/velocityTBH.h
@@ -16,7 +16,7 @@ typedef struct vel_TBH
 	int error;
 	int prevError;
 	bool firstCross;
-	float outValApprox; //Output value at zero error for a given target velocity
+	int outValApprox; //Output value at zero error for a given target velocity
 	float outValAtZero;
 
 	//Timestep
@@ -34,10 +34,10 @@ typedef struct vel_TBH
 } vel_TBH;
 
 //Initializes a velocity TBH controller
-void vel_TBH_InitController(vel_TBH* tbh, tSensors sensor, float gain, float outValApprox);
+void vel_TBH_InitController(vel_TBH* tbh, tSensors sensor, float gain, int outValApprox);
 
 //Sets the target velocity
-void vel_TBH_SetTargetVelocity(vel_TBH* tbh, int targetVelocity, float outValApprox = -1010);
+void vel_TBH_SetTargetVelocity(vel_TBH* tbh, int targetVelocity, int outValApprox = -1010);
 
 //Steps the controller calculations (velocity based)
 int vel_TBH_StepController_VEL(vel_TBH* tbh);
